requirements.txt
Django==5.2.5
python-decouple==3.8
djangorestframework==3.16.0
django-cors-headers==4.7.0
django-extensions==4.1
cloudinary==1.44.1
django-cloudinary-storage==0.3.0
django-rest-knox==5.0.2
drf-yasg==1.21.10
pillow==11.3.0
psycopg2-binary==2.9.10
django-phonenumber-field==8.1.0
phonenumbers==8.12.50
whitenoise==6.9.0
django-filter==25.1